+++ custom-elements-manifest.config.mjs
@@ -1,7 +1,7 @@
 export default {
-  globs: ['src/components/*.ts'],
-  exclude: ['src/components/index.ts'],
-  outdir: '.',
+  globs: ['packages/elements/src/*.ts'],
+  exclude: ['packages/elements/src/index.ts'],
+  outdir: 'packages/elements',
   litelement: true,
   packagejson: false,
 };